refactor(ui): promote DataTable, Pagination, PageLoader to design-system
Move these shared components out of the console app into
@shellhub/design-system/components so other apps can consume them. Each
becomes a named export; DataTable's transitive deps move too (PageLoader)
and it adopts the package's tailwind-merge cn instead of the console's
naive join. Declare the @heroicons/react peer dependency DataTable needs.

Repoint all console importers, relocate the component tests into the
package, and update the firewall page tests' DataTable mock for the
named export. Alert (from the issue) does not exist; ConfirmDialog is
deferred (drags BaseDialog + a11y hooks).

Fixes: shellhub-io/team#151
