+++ ui/apps/console/src/components/ManageTagsDrawer.tsx
@@ -18,7 +18,7 @@ import {
   ExclamationCircleIcon,
 } from "@heroicons/react/24/outline";
 import { Button, IconButton } from "@shellhub/design-system/primitives";
-import PageLoader from "@/components/common/PageLoader";
+import { PageLoader } from "@shellhub/design-system/components";
 
 const TAG_PATTERN = /^[a-zA-Z0-9]+$/;
