+++ ui/apps/console/src/pages/firewall-rules/__tests__/FirewallRules.test.tsx
@@ -60,14 +60,14 @@ vi.mock("@/components/common/ConfirmDialog", () => ({
 
 // Capture DataTable props on every render so we can assert pagination suppression.
 const capturedDataTableProps: Record<string, unknown>[] = [];
-vi.mock("@/components/common/DataTable", async (importOriginal) => {
+vi.mock("@shellhub/design-system/components", async (importOriginal) => {
   const actual =
-    await importOriginal<typeof import("@/components/common/DataTable")>();
+    await importOriginal<typeof import("@shellhub/design-system/components")>();
   return {
     ...actual,
-    default: (props: Record<string, unknown>) => {
+    DataTable: (props: Record<string, unknown>) => {
       capturedDataTableProps.push({ ...props });
-      return actual.default(props as unknown as Parameters<typeof actual.default>[0]);
+      return actual.DataTable(props as unknown as Parameters<typeof actual.DataTable>[0]);
     },
   };
 });
