+++ ui/packages/design-system/__tests__/Pagination.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from "vitest";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import Pagination from "../Pagination";
+import { Pagination } from "../components/Pagination";
 
 describe("Pagination", () => {
   it("Prev and Next buttons have explicit type='button'", () => {
