+++ ui/packages/design-system/components/Pagination.tsx
@@ -6,7 +6,7 @@ interface Props {
   onPageChange: (page: number) => void;
 }
 
-export default function Pagination({
+export function Pagination({
   page,
   totalPages,
   totalCount,
