+++ ui/packages/design-system/components/index.ts
@@ -3,3 +3,7 @@ export { CopyButton } from "./CopyButton";
 export { ConnectionGrid } from "./ConnectionGrid";
 export { ShimmerCard } from "./ShimmerCard";
 export { Footer } from "./Footer";
+export { Pagination } from "./Pagination";
+export { PageLoader } from "./PageLoader";
+export { DataTable } from "./DataTable";
+export type { Column, DataTableProps } from "./DataTable";
