+++ ui/packages/design-system/package.json
@@ -20,6 +20,7 @@
     "tailwind-merge": "^2.6.0"
   },
   "peerDependencies": {
+    "@heroicons/react": "^2.2.0",
     "react": "^18 || ^19"
   }
 }
